improvement of categories and calendar location
the category input has been replaced with a dynamic select with custom colours and visual category management has been added. A button has been implemented to add the current location via geolocation, and the location usage flow has been improved in both the calendar and the centre search. The event details modal has been redesigned with a clearer layout and contextual actions. Styles have been updated for a better user experience, and the management of the deletion of associated categories and events has been enhanced.